requirements.txt
fastapi[standard]
uvicorn[standard]
sqlmodel 
sqlalchemy[asyncio]
asyncpg 
pydantic
python-dotenv
mistralai
pytest
python-dotenv
datauri
openai
Jinja2
pydantic_settings